+++ docs/search/all_a.js
@@ -1,8 +1,8 @@
 var searchData=
 [
   ['lasttime_0',['lastTime',['../struct_c_s_c_i441_1_1_m_d5_model_1_1_m_d5_animation_state.html#adf49659d1d428a4e505ef70e11e8e119',1,'CSCI441::MD5Model::MD5AnimationState']]],
-  ['loadandregister2dtexture_1',['loadAndRegister2DTexture',['../_texture_utils_8hpp.html#af423a3674d7787d7ca100213acea5a3e',1,'CSCI441::TextureUtils']]],
-  ['loadandregistertexture_2',['loadAndRegisterTexture',['../_texture_utils_8hpp.html#a357b87b32ddd807943848055136e94d0',1,'CSCI441::TextureUtils']]],
+  ['loadandregister2dtexture_1',['loadAndRegister2DTexture',['../_texture_utils_8hpp.html#a414ebb3c74d5281a5d4f545a9360d9ff',1,'CSCI441::TextureUtils']]],
+  ['loadandregistertexture_2',['loadAndRegisterTexture',['../_texture_utils_8hpp.html#a7760a68a85880637fe696d1ef0a2f19a',1,'CSCI441::TextureUtils']]],
   ['loadcubemapfacetexture_3',['loadCubeMapFaceTexture',['../_texture_utils_8hpp.html#a1c32bc9d6e247f7b8a9b8b8b23b605d1',1,'CSCI441::TextureUtils']]],
   ['loadmd5model_4',['loadMD5Model',['../class_c_s_c_i441_1_1_m_d5_model.html#a57f46990ad43586e863846633803ac70',1,'CSCI441::MD5Model']]],
   ['loadmodelfile_5',['loadModelFile',['../class_c_s_c_i441_1_1_model_loader.html#a6a41de5f070d8cf0964e47729c1d8482',1,'CSCI441::ModelLoader']]],